LabExamples/Lab08_Animasyon_Zamanlayicilar/pyproject.toml
[project]
name = "lab08-animasyon-zamanlayicilar"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.7"]
